Added a helper_functions file to store functions used by multiple files. Updated the upload api URL to follow a convention. Added an api file to store api functions.
